+++ src/core_modules/capture-core/components/DataEntry/dataEntryOutput/withFeedbackOutput.tsx
@@ -3,13 +3,14 @@ import { connect } from 'react-redux';
 import i18n from '@dhis2/d2-i18n';
 import { getDataEntryKey } from '../common/getDataEntryKey';
 import { withDataEntryOutput } from './withDataEntryOutput';
-import { WidgetFeedback } from '../../WidgetFeedback';
-import type { FilteredText, FilteredKeyValue } from '../../WidgetFeedback';
+import {
+    WidgetFeedback,
+    type FilteredFeedbackText, type FilteredFeedbackKeyValue } from '../../WidgetFeedback';
 
 type Props = {
     feedbackItems: {
-        displayTexts: Array<FilteredText>;
-        displayKeyValuePairs: Array<FilteredKeyValue>;
+        displayTexts: Array<FilteredFeedbackText>;
+        displayKeyValuePairs: Array<FilteredFeedbackKeyValue>;
     };
 };
 
@@ -28,7 +29,7 @@ const getFeedbackOutput = () =>
             return (
                 <div>
                     {hasItems &&
-                        <WidgetFeedback feedback={feedback} emptyText={i18n.t('No feedback for this event yet')} />
+                        <WidgetFeedback feedback={feedback} feedbackEmptyText={i18n.t('No feedback for this event yet')} />
                     }
                 </div>
             );