+++ src/core_modules/capture-core/components/WidgetProfile/DataEntry/hooks/useProgramRules.ts
@@ -3,8 +3,8 @@ import { useDataQuery } from '@dhis2/app-runtime';
 
 const fields =
     'id,displayName,condition,description,program[id],programStage[id],priority,' +
-    'programRuleActions[id,content,displayContent,location,data,programRuleActionType,programStageSection[id],dataElement[id],' +
-    'trackedEntityAttribute[id],programStage[id],optionGroup[id],option[id]]';
+    'programRuleActions[id,content,displayContent,location,data,programRuleActionType,' +
+    'programStageSection[id],dataElement[id],trackedEntityAttribute[id],programStage[id],optionGroup[id],option[id]]';
 
 export const useProgramRules = (programId: string) => {
     const [page, setPage] = useState(0);