+++ src/core_modules/capture-core/components/WorkingLists/TrackerWorkingLists/TrackedEntityBulkActions/TrackedEntityBulkActions.container.tsx
@@ -19,7 +19,8 @@ export const TrackedEntityBulkActions = ({
     const { activeList } = useBulkDataEntryConfigurations(programId);
 
     const injectSelectedRowsToBulkDataEntryPlugin = useCallback(() => {
-        recordsOrder && onOpenBulkDataEntryPlugin && onOpenBulkDataEntryPlugin(recordsOrder.filter(recordOrder => selectedRows[recordOrder]));
+        recordsOrder && onOpenBulkDataEntryPlugin &&
+            onOpenBulkDataEntryPlugin(recordsOrder.filter(recordOrder => selectedRows[recordOrder]));
     }, [onOpenBulkDataEntryPlugin, recordsOrder, selectedRows]);
 
     if (programStageId) {