+++ src/core_modules/capture-core/components/WorkingLists/TrackerWorkingLists/epics/teiViewEpics/helpers/initTrackerWorkingListsView.types.ts
@@ -1,4 +1,8 @@
-import type { TeiColumnsMetaForDataFetching, TeiFiltersOnlyMetaForDataFetching, TrackerWorkingListsTemplate } from '../../../types';
+import type {
+    TeiColumnsMetaForDataFetching,
+    TeiFiltersOnlyMetaForDataFetching,
+    TrackerWorkingListsTemplate,
+} from '../../../types';
 import type { QuerySingleResource } from '../../../../../../utils/api';
 
 export type Input = {